# streamlit/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SuperPandas",
        "description": "A production-ready FastAPI template with LangGraph and Langfuse integration",
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/v1/chatbot/select-agent": {
            "post": {
                "tags": [
                    "chatbot"
                ],
                "summary": "Select Agent",
                "description": "Select an agent for the current session.\n\nArgs:\n    request: The FastAPI request object for rate limiting.\n    agent_selection: The agent selection request.\n    session: The current session from the auth token.\n\nReturns:\n    dict: A message indicating the agent was selected.\n\nRaises:\n    HTTPException: If the agent selection is invalid.",
                "operationId": "select_agent_api_v1_chatbot_select_agent_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AgentSelection"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/chatbot/chat": {
            "post": {
                "tags": [
                    "chatbot"
                ],
                "summary": "Chat",
                "description": "Process a chat request using LangGraph.\n\nArgs:\n    request: The FastAPI request object for rate limiting.\n    chat_request: The chat request containing messages.\n    session: The current session from the auth token.\n\nReturns:\n    ChatResponse: The processed chat response.\n\nRaises:\n    HTTPException: If there's an error processing the request.",
                "operationId": "chat_api_v1_chatbot_chat_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChatRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChatResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/chatbot/chat/stream": {
            "post": {
                "tags": [
                    "chatbot"
                ],
                "summary": "Chat Stream",
                "description": "Process a chat request using LangGraph with streaming response.\n\nArgs:\n    request: The FastAPI request object for rate limiting.\n    chat_request: The chat request containing messages.\n    session: The current session from the auth token.\n\nReturns:\n    StreamingResponse: A streaming response of the chat completion.\n\nRaises:\n    HTTPException: If there's an error processing the request.",
                "operationId": "chat_stream_api_v1_chatbot_chat_stream_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChatRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/chatbot/messages": {
            "get": {
                "tags": [
                    "chatbot"
                ],
                "summary": "Get Session Messages",
                "description": "Get all messages for a session.\n\nArgs:\n    request: The FastAPI request object for rate limiting.\n    session: The current session from the auth token.\n\nReturns:\n    ChatResponse: All messages in the session.\n\nRaises:\n    HTTPException: If there's an error retrieving the messages.",
                "operationId": "get_session_messages_api_v1_chatbot_messages_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChatResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "chatbot"
                ],
                "summary": "Clear Chat History",
                "description": "Clear all messages for a session.\n\nArgs:\n    request: The FastAPI request object for rate limiting.\n    session: The current session from the auth token.\n\nReturns:\n    dict: A message indicating the chat history was cleared.",
                "operationId": "clear_chat_history_api_v1_chatbot_messages_delete",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AgentSelection": {
                "properties": {
                    "platform": {
                        "type": "string",
                        "enum": [
                            "fic",
                            "sevdesk",
                            "hr"
                        ],
                        "title": "Platform",
                        "description": "The platform to select for the session"
                    }
                },
                "type": "object",
                "required": [
                    "platform"
                ],
                "title": "AgentSelection",
                "description": "Request model for agent selection endpoint.\n\nAttributes:\n    platform: The platform to select for the session."
            },
...
            "ChatRequest": {
                "properties": {
                    "messages": {
                        "items": {
                            "$ref": "#/components/schemas/Message"
                        },
                        "type": "array",
                        "minItems": 1,
                        "title": "Messages",
                        "description": "List of messages in the conversation"
                    },
                    "platform": {
                        "anyOf": [
                            {
                                "type": "string",
                                "enum": [
                                    "fic",
                                    "sevdesk",
                                    "hr"
                                ]
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Platform",
                        "description": "The platform to use for the chat"
                    }
                },
                "type": "object",
                "required": [
                    "messages"
                ],
                "title": "ChatRequest",
                "description": "Request model for chat endpoint.\n\nAttributes:\n    messages: List of messages in the conversation."
            },
            "ChatResponse": {
                "properties": {
                    "messages": {
                        "items": {
                            "$ref": "#/components/schemas/Message"
                        },
                        "type": "array",
                        "title": "Messages",
                        "description": "List of messages in the conversation"
                    },
                    "generated_code": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Generated Code",
                        "description": "The generated code to answer the user's question."
                    },
                    "fig": {
                        "anyOf": [
                            {},
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Fig",
                        "description": "The figure for plotting the data."
                    }
                },
                "type": "object",
                "required": [
                    "messages"
                ],
                "title": "ChatResponse",
                "description": "Response model for chat endpoint.\n\nAttributes:\n    messages: List of messages in the conversation.\n    generated_code: The generated code to answer the user's question.\n    fig: The figure for plotting the data."
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Message": {
                "properties": {
                    "role": {
                        "type": "string",
                        "enum": [
                            "user",
                            "assistant",
                            "system"
                        ],
                        "title": "Role",
                        "description": "The role of the message sender"
                    },
                    "content": {
                        "type": "string",
                        "maxLength": 12000,
                        "minLength": 1,
                        "title": "Content",
                        "description": "The content of the message"
                    }
                },
                "type": "object",
                "required": [
                    "role",
                    "content"
                ],
                "title": "Message",
                "description": "Message model for chat endpoint.\n\nAttributes:\n    role: The role of the message sender (user or assistant).\n    content: The content of the message."
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}